Move dynamic preload code to package
